getExternalMountsを新しく作ったStorageUtilsへ移動
